[cacache] Update to 15.0
