+++ types/cacache/cacache-tests.ts
@@ -9,7 +9,7 @@ cacache.ls.stream(cachePath).on('data', data => {
     data; // $ExpectType any
 });
 
-cacache.get(cachePath, 'my-thing').then(() => {});
+cacache.get(cachePath, 'my-thing', { memoize: true }).then(() => {});
 
 cacache.get.byDigest(cachePath, 'sha512-BaSe64HaSh').then(() => {});
 
@@ -29,14 +29,19 @@ cacache.get.info(cachePath, 'my-thing').then(() => {});
 
 cacache.get.hasContent(cachePath, 'sha521-NOT+IN/CACHE==').then(() => {});
 
-cacache.put(cachePath, 'registry.npmjs.org|cacache@1.0.0', Buffer.from([])).then(integrity => {
-    integrity; // $ExpectType string
-});
+cacache
+    .put(cachePath, 'registry.npmjs.org|cacache@1.0.0', Buffer.from([]), {
+        tmpPrefix: 'tmp-',
+        memoize: true,
+    })
+    .then(integrity => {
+        integrity; // $ExpectType string
+    });
 
 fs.createReadStream('').pipe(
-  cacache.put
-    .stream(cachePath, 'registry.npmjs.org|cacache@1.0.0')
-    .on('integrity', d => console.log(`integrity digest is ${d}`)),
+    cacache.put
+        .stream(cachePath, 'registry.npmjs.org|cacache@1.0.0')
+        .on('integrity', d => console.log(`integrity digest is ${d}`)),
 );
 
 cacache.rm.all(cachePath).then(() => {
@@ -51,25 +56,54 @@ cacache.rm.content(cachePath, 'sha512-SoMeDIGest/IN+BaSE64==').then(() => {
     console.log('data for my-thing is gone!');
 });
 
-cacache.tmp.mkdir(cachePath).then(dir => {
-    dir; // $ExpectType string
-    // ...
-});
-
-cacache.tmp.mkdir(cachePath).then(dir => {
-    dir; // $ExpectType string
-    cacache.tmp.fix(cachePath).then(() => {
+cacache.tmp
+    .mkdir(cachePath, {
+        concurrency: 10,
+        filter: false,
+        log: { silly: () => {} },
+        tmpPrefix: 'tmp-',
+    })
+    .then(dir => {
+        dir; // $ExpectType string
         // ...
     });
-});
 
-cacache.tmp.withTmp(cachePath, dir => {
-    dir; // $ExpectType string
-});
+cacache.tmp
+    .mkdir(cachePath, {
+        concurrency: 10,
+        filter: false,
+        log: { silly: () => {} },
+        tmpPrefix: 'tmp-',
+    })
+    .then(dir => {
+        dir; // $ExpectType string
+        cacache.tmp.fix(cachePath).then(() => {
+            // ...
+        });
+    });
 
-cacache.verify(cachePath).then(stats => {
-    console.log('cache is much nicer now! stats:', stats);
-});
+cacache.tmp.withTmp(
+    cachePath,
+    {
+        concurrency: 10,
+        filter: false,
+        log: { silly: () => {} },
+        tmpPrefix: 'tmp-',
+    },
+    dir => {
+        dir; // $ExpectType string
+    },
+);
+
+cacache
+    .verify(cachePath, {
+        concurrency: 10,
+        filter: false,
+        log: { silly: () => {} },
+    })
+    .then(stats => {
+        console.log('cache is much nicer now! stats:', stats);
+    });
 
 cacache.verify(cachePath).then(() => {
     cacache.verify.lastRun(cachePath).then(lastTime => {
